Collapse Types down for Images, and remove type aliases.
