+++ src/lib/Pixel/Error.hs
@@ -1,11 +1,8 @@
 module Pixel.Error
   ( Error(..)
-  , ImageError(..)
-  )
-where
+  ) where
 
 import Protolude
-import Pixel.Model.Images.Error ( ImageError(..) )
 import Pixel.Model.Users.Error  ( UserError(..) )
 
 --------------------------------------------------------------------------------
@@ -17,7 +14,6 @@ data AuthenticationError
 
 data Error
   = AuthError  !AuthenticationError
-  | ImageError !ImageError
   | UserError  !UserError
   | UnknownError
   deriving Show
