+++ src/lib/Pixel/Lens.hs
@@ -2,7 +2,7 @@ module Pixel.Lens where
 
 import Control.Lens
 import Pixel.API.Types
-import Pixel.Model.Images.Types
+import Pixel.Model.Images
 import Pixel.Model.Users.Types.Email
 import Pixel.Model.Users.Types.Password
 import Pixel.Model.Users.Types.User
