--- src/lib/Pixel/Model/Images/Error.hs
@@ -1,12 +1,0 @@
-module Pixel.Model.Images.Error ( ImageError(..) ) where
-
-import Protolude
-
---------------------------------------------------------------------------------
-
-data ImageError
-  = MissingToken
-  | InvalidUUID
-  | ImageDoesNotExist
-  deriving Show
-
